+++ src/client-state.ts
@@ -12,12 +12,6 @@ export type ClientState = Entity & {
   yjsAwarenessState: JSONObject;
 };
 
-export type UserInfo = {
-  name: string;
-  avatar: string;
-  color: string;
-};
-
 const {
   get: getClientState,
   delete: deleteClientState,
@@ -56,28 +50,4 @@ export async function yjsSetLocalState(
   } else {
     await putClientState(tx, {id: tx.clientID, yjsClientID, yjsAwarenessState});
   }
-}
-
-export function randUserInfo(): UserInfo {
-  const [avatar, name] = avatars[randInt(0, avatars.length - 1)];
-  return {
-    avatar,
-    name,
-    color: colors[randInt(0, colors.length - 1)],
-  };
-}
-
-const colors = ['#f94144', '#f3722c', '#f8961e', '#f9844a', '#f9c74f'];
-const avatars = [
-  ['🐶', 'Puppy'],
-  ['🐱', 'Kitty'],
-  ['🐭', 'Mouse'],
-  ['🐹', 'Hamster'],
-  ['🐰', 'Bunny'],
-];
-
-function randInt(min: number, max: number) {
-  min = Math.ceil(min);
-  max = Math.floor(max);
-  return Math.floor(Math.random() * (max - min + 1) + min); //The maximum is inclusive and the minimum is inclusive
 }
